2.3-aiohttp/requirements.txt
aiohttp
aiohttp[speedups]
aiohttp_basicauth
sqlalchemy
asyncpg
python-dotenv
passlib
pydantic
bcrypt
